+++ astropy/io/fits/tests/__init__.py
@@ -25,13 +25,7 @@ class FitsTestCase(object):
         fits.STRIP_HEADER_WHITESPACE.set(True)
         fits.USE_MEMMAP.set(True)
 
-        # Ignore deprecation warnings--this only affects Python 2.5 and 2.6,
-        # since deprecation warnings are ignored by defualt on 2.7
-        warnings.simplefilter('ignore')
-        warnings.simplefilter('always', UserWarning)
-
     def teardown(self):
-        warnings.resetwarnings()
         if hasattr(self, 'temp_dir') and os.path.exists(self.temp_dir):
             tries = 3
             while tries: