+++ astropy/io/fits/tests/util.py
@@ -3,8 +3,10 @@
 
 import warnings
 
+from ....tests.helper import catch_warnings
 
-class ignore_warnings(warnings.catch_warnings):
+
+class ignore_warnings(catch_warnings):
     def __enter__(self):
         retval = super(ignore_warnings, self).__enter__()
         warnings.simplefilter('ignore')