+++ CHANGES.rst
@@ -341,6 +341,9 @@ Bug Fixes
     on big endian platforms, such as MIPS, PPC, s390x and
     SPARC. [#1849]
 
+  - The test suite will now raise exceptions when a deprecated feature
+    of Python or Numpy is used.  [#1948]
+
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^